RGB LED
The code still has all previous functions implemented. Now an additional RGB LED is controlled by the encoder. One color is changed at a time in the same way that the original LED was, except it isn't set to 0 on a button press. Now if the button is pressed it will switch which RGB value is being controlled. The RGB control state is shown in the output Digital LEDs of their respective color